refactor(LottoGame)
