+++ src/main/java/lotto/domain/numbers/LottoTickets.java
@@ -1,5 +1,8 @@
 package lotto.domain.numbers;
 
+import lotto.domain.LottoResult;
+import lotto.domain.Rank;
+
 import java.util.List;
 
 public class LottoTickets {
@@ -11,6 +14,15 @@ public class LottoTickets {
 
     public int getCount() {
         return lottoTickets.size();
+    }
+
+    public LottoResult getMatchLottoResult(final WinningLottoTicket winningLottoTicket) {
+        LottoResult lottoResult = new LottoResult();
+        lottoTickets.forEach(lottoTicket -> {
+            Rank rank = lottoTicket.matchWinningLottoNumbers(winningLottoTicket);
+            lottoResult.addRank(rank);
+        });
+        return lottoResult;
     }
 
     public List<LottoTicket> getLottoTickets() {
